Move CharStats out of MPQInit
