+++ BH/MPQInit.cpp
@@ -9,7 +9,7 @@ bool initialized = false;
 
 std::vector<StatProperties*> AllStatList;
 std::unordered_map<std::string, StatProperties*> StatMap;
-std::vector<CharStats*> CharList;
+
 std::map<std::string, ItemAttributes*> ItemAttributeMap;
 std::map<std::string, InventoryLayout*> InventoryLayoutMap;
 std::vector<ItemAffixProperties*> AllPrefixList;
@@ -1099,16 +1099,6 @@ void InitializeMPQData() {
 
 	char* end;
 	short lastID = -1;
-
-	if (MpqDataMap.find("charstats") != MpqDataMap.end()) {
-		for (auto d = MpqDataMap["charstats"]->data.begin(); d < MpqDataMap["charstats"]->data.end(); d++) {
-			if ((*d)["ToHitFactor"].length() > 0) {
-				CharStats* bits = new CharStats();
-				bits->toHitFactor = std::stoi((*d)["ToHitFactor"].c_str(), nullptr, 10);
-				CharList.push_back(bits);
-			}
-		}
-	}
 
 	if (MpqDataMap.find("itemstatcost") != MpqDataMap.end()) {
 		for (auto d = MpqDataMap["itemstatcost"]->data.begin(); d < MpqDataMap["itemstatcost"]->data.end(); d++) {
