+++ BH/MPQInit.h
@@ -61,13 +61,11 @@ struct ItemAffixProperties {
 	unsigned int costMultiply;
 };
 
-struct CharStats {
-	int toHitFactor;
-};
+
 
 extern std::vector<StatProperties*> AllStatList;
 extern std::unordered_map<std::string, StatProperties*> StatMap;
-extern std::vector<CharStats*> CharList;
+
 extern std::map<std::string, ItemAttributes*> ItemAttributeMap;
 extern std::map<std::string, InventoryLayout*> InventoryLayoutMap;
 extern std::vector<ItemAffixProperties*> AllPrefixList;
